pyproject.toml
[tool.poetry]
name = "reviewturtl"
version = "0.1.0"
description = "Your Open Source Review tool!"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.4"
fastapi = "^0.111.0"
dspy-ai = "^2.4.12"
inflection = "^0.5.1"
fastembed = "0.3.4"
qdrant-client = "^1.10.1"
sentence-transformers = "^3.0.1"
ruff = "^0.5.3"
uvicorn = "^0.30.6"
prisma = "0.15.0"
cachetools = "^5.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
